Set up healer actor with health on UI

+++ Source/Carpocalypse/Private/EnemyHealth.cpp
@@ -23,9 +23,9 @@ void AEnemyHealth::Tick(float DeltaTime)
 
 }
 
-int AEnemyHealth::SetHealth(EEnemyTypes enemyType)
+float AEnemyHealth::SetHealth(EEnemyTypes enemyType)
 {
-	int health;
+	float health;
 	
 	switch (enemyType)
 	{
